requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
boto3==1.6.23
botocore==1.9.23          # via boto3, s3transfer
dj-database-url==0.5.0
django-enumchoicefield==1.0.0
django-filter==1.1.0
django-js-asset==1.1.0    # via django-mptt
django-mptt==0.9.0
django-positions==0.6.0
django-storages==1.6.6
django==2.0.5
djangorestframework==3.8.2
docutils==0.14            # via botocore
feedparser==5.2.1
jmespath==0.9.3           # via boto3, botocore
psycopg2==2.7.4
python-dateutil==2.6.1    # via botocore
pytz==2018.4              # via django
s3transfer==0.1.13        # via boto3
six==1.11.0               # via python-dateutil
